Update for week1

+++ Search in Rotated Sorted Array/Search in Rotated Sorted Array/main.cpp
@@ -11,7 +11,7 @@ class Solution {
 public:
     int search(vector<int>& nums, int target) {
         int s = 0;
-        int e = nums.size() - 1;
+        int e = (int)nums.size() - 1;
         
         while (s <= e)
         {
